requirements.txt
requests
pyyaml
dnsgen
pymongo
tldextract
python-nmap
boto3
pytest
